+++ examples/simple_repeater/MyMesh.h
@@ -123,7 +123,7 @@ class MyMesh : public mesh::Mesh, public CommonCLICallbacks
   CayenneLPP telemetry;
   unsigned long set_radio_at, revert_radio_at;
   unsigned long _ota_update_at = 0;  // deferred `ota update` fire time (0 = none scheduled)
-#if defined(WITH_MQTT_BRIDGE) && defined(WITH_QTA_PRIVATE_BOT)
+#ifdef WITH_MQTT_BRIDGE
   unsigned long _qta_bot_next_reply_ms = 0;
 #endif
   float pending_freq;
@@ -271,6 +271,8 @@ protected:
   void onControlDataRecv(mesh::Packet* packet) override;
   int searchChannelsByHash(const uint8_t* hash, mesh::GroupChannel channels[], int max_matches) override;
   void onGroupDataRecv(mesh::Packet* packet, uint8_t type, const mesh::GroupChannel& channel, uint8_t* data, size_t len) override;
+  bool isBotPublicChannel(const mesh::GroupChannel& channel) const;
+  bool sendBotReply(mesh::GroupChannel& channel, const char* text);
 
   void sendFloodReply(mesh::Packet* packet, unsigned long delay_millis, uint8_t path_hash_size);